remove communication.rs, move contents to encryption.rs
